Feature/lesson12

+++ lesson10/src/main/java/com/andersenlab/spring/service/impl/UserServiceImpl.java
@@ -1,7 +1,8 @@
 package com.andersenlab.spring.service.impl;
 
+import com.andersenlab.spring.dto.UserDto;
+import com.andersenlab.spring.dto.dtoMapper.UserDtoMapper;
 import com.andersenlab.spring.entity.User;
-import com.andersenlab.spring.repository.TicketRepository;
 import com.andersenlab.spring.repository.UserRepository;
 import com.andersenlab.spring.service.UserService;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -10,46 +11,70 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
-    private final TicketRepository ticketRepository;
 
-    public UserServiceImpl(UserRepository userRepository, TicketRepository ticketRepository) {
+    public UserServiceImpl(UserRepository userRepository) {
         this.userRepository = userRepository;
-        this.ticketRepository = ticketRepository;
     }
 
     @Override
-    public User saveUser(User user) {
-        return userRepository.save(user);
+    public UserDto saveUser(UserDto user) {
+        userRepository.save(UserDtoMapper.toUser(user));
+        return user;
     }
 
     @Override
-    public User getUserById(int id) {
-        return userRepository.findById(id).orElse(null);
+    public UserDto getUserById(int id) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            return UserDtoMapper.toUserDto(user.get());
+        }
+        throw new IllegalArgumentException("User not found");
     }
 
     @Override
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public List<UserDto> getAllUsers() {
+        List<User> users = userRepository.findAll();
+        List<UserDto> userDtos = new ArrayList<>();
+        if (!users.isEmpty()) {
+            for (User user : users) {
+                userDtos.add(UserDtoMapper.toUserDto(user));
+            }
+            return userDtos;
+        }
+        throw new IllegalArgumentException("Tickets not found");
     }
 
     @Override
     public void deleteUser(int id) {
-        ticketRepository.findByUserId(id).forEach(x -> ticketRepository.deleteById(x.getId()));
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            System.out.println("User deleted successfully");
+        } else throw new IllegalArgumentException("User not found");
     }
 
     @Override
-    public User updateUserName(int id, String name) {
-        User newUser = userRepository.findById(id).orElse(null);
-        if (newUser != null) {
-            newUser.setName(name);
-            return userRepository.save(newUser);
+    public UserDto updateUserName(int id, String name) {
+        Optional<User> newUser = userRepository.findById(id);
+        if (newUser.isPresent()) {
+            newUser.get().setName(name);
+            return UserDtoMapper.toUserDto(userRepository.save(newUser.get()));
+        }
+        throw new IllegalArgumentException("User not found");
+    }
+
+    public UserDto updateUserStatus(int id) {
+        Optional<User> newUser = userRepository.findById(id);
+        if (newUser.isPresent()) {
+            newUser.get().setUserStatus(true);
+            return UserDtoMapper.toUserDto(userRepository.save(newUser.get()));
         }
         throw new IllegalArgumentException("User not found");
     }
